feat: Add Help component
This commit adds a new file, Help.tsx, which contains the implementation of the Help component. The Help component will provide assistance and guidance to users.
feat: Add Help component

This commit adds a new file, Help.tsx, which contains the implementation of the Help component. The Help component will provide assistance and guidance to users.